+++ i18n/zh-CN/readme/CHANGELOG.md
@@ -2,8 +2,8 @@
 
     + Added shortcut for pushing reference to TeXstudio with Better BibTeX
     + Add "Advanced search" shortcut
-    + Add "Locate with Google Scholar" shortcut
-    + Add "Locate menu" shortcut
+    + Add "Locate item" shortcut
+    + Add "Show Locate menu" shortcut
 
 * In version 3.3.3: